% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusion.exclusion.R
\name{apply.inc.exc}
\alias{apply.inc.exc}
\title{Using parsed control inclusion/exclusion data,
remove controls that don't match specified criteria.}
\usage{
apply.inc.exc(
  df,
  phenotype.name,
  trait.is.binary,
  control.inclusion.list,
  control.exclusion.list
)
}
\arguments{
\item{df}{data.frame, input model matrix data}

\item{phenotype.name}{character vector, a column
name from \code{df}}

\item{trait.is.binary}{logical, an indicator whether
the phenotype is binary (0/1, 0 == control)}

\item{control.inclusion.list}{list, inclusion criteria
for controls in terms of other categorical variables from
\code{df}, as processed by \code{load.inc.exc}}

\item{control.exclusion.list}{list, exclusion criteria
for controls in terms of other categorical variables from
\code{df}, as processed by \code{load.inc.exc}}
}
\description{
This function takes an under-construction data frame
of model matrix information, a phenotype, and control
inclusion/exclusion lists from \code{load.inc.exc}, and
applies the filters requested by the inclusion/exclusion
lists. Filters only impact controls, and by extension
this function does nothing at all if provided with
a trait that isn't binary.
}
