% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.normalization.R
\name{apply.inverse.normalization}
\alias{apply.inverse.normalization}
\title{Apply inverse normalization to selected columns of a data frame}
\usage{
apply.inverse.normalization(df, conditions)
}
\arguments{
\item{df}{data.frame, source data frame containing model matrix
data to be transformed}

\item{conditions}{list of logical vectors. each vector should
be of length \code{ncol(df)}, and should indicate whether a column
passes a particular criterion for being included in the
transformed set. each vector is effectively logically ORed together.}
}
\description{
This function takes a data frame, constructed of
model matrix data from a phenotype dataset, and applies
the inverse normal transformation to each column matching
a certain set of provided conditions. The variables are
stratified by sex before application of the transformation.
An extension here would be to allow arbitrary stratification
variables before transformation.
}
