% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.model.matrix.R
\name{construct.model.matrix}
\alias{construct.model.matrix}
\title{Given configuration and phenotype data
create a model matrix.}
\usage{
construct.model.matrix()
}
\description{
This function is currently a placeholder.
}
