% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.handling.R
\name{extract.categories}
\alias{extract.categories}
\title{Read category tracker and extract categorical levels}
\usage{
extract.categories(df, phenotype.name, category.filename)
}
\arguments{
\item{df}{data.frame, working data frame with phenotype
variable present for all relevant subjects, plus possibly
some extra rows or columns}

\item{phenotype.name}{character vector, column name of the
phenotype variable in \code{df}}

\item{category.filename}{character vector, name of
preprocessor *.categories tracker file with rows
of category levels and either "reference" or "comparison",
delimited by tabs}
}
\description{
This function takes the preprocessor *.categories file,
loads its contents, and then uses its annotations to
merge/filter categorical levels until a single binary
comparison is left. This binary comparison is what is
actually run through SAIGE for categorical traits.
}
\seealso{
construct.model.matrix
}
