% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.normalization.R
\name{inverse.normalize}
\alias{inverse.normalize}
\title{Apply inverse normalization to continuous variables}
\usage{
inverse.normalize(i)
}
\arguments{
\item{i}{numeric vector, values to be transformed}
}
\description{
Performs the inverse (rank) normal transformation
on a vector of values. You can do this to binary
data but you're not gonna like the results lol.
}
