% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.loading.R
\name{load.inc.exc}
\alias{load.inc.exc}
\title{Load control inclusion or exclusion data from file}
\usage{
load.inc.exc(filename)
}
\arguments{
\item{filename}{character vector, name of a file
containing control inclusion or exclusion data, or the
string "NA". format for this
file is: per row, a variable name, and optionally
a comma-delimited list of variable categories denoting
valid controls. for backwards compatibility, a variant
of this file only containing the first column is permitted,
in which case all non-zero levels of the variable will be
considered inclusion levels.}
}
\description{
This function loads several variants of control inclusion
or exclusion data from file into a list of variable data.
Inclusions and exclusions are specified in the same format
and thus handled identically by this function.
}
